Revert "fix: PAPI and Paraspell Integration"
This reverts commit d7bb497e20a3ea04b2f748b1c3c510748bb8b558.
